pyproject.toml
[tool.poetry]
name = "unboundai-gtm-mcp"
version = "0.0.5"
description = "MCP server for Google Tag Manager"
authors = ["Paolo Bietolini", "Tije Vlam"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/tijevlam/gtm-mcp"
repository = "https://github.com/tijevlam/gtm-mcp"
documentation = "https://github.com/tijevlam/gtm-mcp#readme"
keywords = ["mcp", "google-tag-manager", "gtm", "model-context-protocol", "claude", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
packages = [{ include = "unboundai_gtm_mcp", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
google-api-python-client = "^2.186.0"
google-auth = "^2.41.0"
mcp = "^1.20.0"
protobuf = "^6.33.0"
python-dotenv = "^1.2.1"

[tool.poetry.scripts]
unboundai-gtm-mcp = "unboundai_gtm_mcp.server:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
